Initial commit for the Galaxy S 4G (SGH-T959V)
This commit gets close to getting things working.

There are two things strange in this commit.

1) Nand is recognized, but not scan-able.
2) While attempting to boot to recovery, /init is not found.

I have logged and provided the usb-uart log of the kernel here:
	Documentation/SGH-T959V/initial-commit.log

A note on #1:
 When I initially worked on getting GB MTD to work on this board, I
had to use the samsung_audi.c driver that epic4g used from GT-I9100.

I checked the ICS drop of GT-I9100, and it is very close to the aries
provided drivers, so I don't see what is different. I need to
investigate this more.

Change-Id: I23643c7036087968250aeba803602d11d3e1d7eb